move integration tests into acquisition
